This PR makes the cli to be in line with argparse implementation in terms of startup time. Since we register all subcommands by import in `__main__` trying to use one command might be slow due to unrelated command being slow as it does heavy import. For example using sync-perm might be slow because of webserver that imports `create_app` from `www`. The PR moves the imports to local imports so that the imports are done as needed. There is some code duplication due to local imports but it's worth the speedup.

Regarding `db` command the default value shown has `config_dict` imported and sorted at module level making the import of the file slow. I first tried making default as a callable but click just shows `dynamic` when a function is used. So I have resorted to using a class that emulates a callable to skip check and on `__str__` it returns the sorted tables. This ensures the `db cleanup --help` only evaluates the import.

A question I had was that in argparse the list is shown in help but no default is set meanwhile in click implementation the default is made as all_tables which means by default it cleans up all tables I suppose that might lead to surprising result compared to argparse.

| implementation  | startup time (seconds) | imports  |
|---|---|---|
|  argparse | 0.686  |  614 |
|  current | 2.327 | 2763  |
|  PR | 0.671 |  641 |

Argparse cli (0.686 seconds, 614 imports)

```
(.env) ➜  airflow git:(fix-imports) time python -Wignore -m airflow --help 2>&1 > /dev/null
python -Wignore -m airflow --help 2>&1 > /dev/null  0.64s user 0.04s system 99% cpu 0.686 total

(.env) ➜  airflow git:(convert-to-click-cli) python -X importtime -m airflow 2>&1 | grep 'import time' | wc   
    614    4300   40655
```

Before (2.327 seconds, 2763 imports)

```
(.env) ➜  airflow git:(convert-to-click-cli) time python -Wignore -m airflow.cli --help 2>&1 > /dev/null
python -Wignore -m airflow.cli --help 2>&1 > /dev/null  2.48s user 1.21s system 158% cpu 2.327 total

(.env) ➜  airflow git:(convert-to-click-cli) python -X importtime -m airflow.cli 2>&1 | grep 'import time' | wc
   2763   19343  230084
```

After (0.671 seconds, 641 imports)

```
(.env) ➜  airflow git:(fix-imports) time python -Wignore -m airflow.cli --help 2>&1 > /dev/null
python -Wignore -m airflow.cli --help 2>&1 > /dev/null  0.63s user 0.04s system 98% cpu 0.671 total

(.env) ➜  airflow git:(fix-imports) python -X importtime -m airflow.cli 2>&1 | grep 'import time' | wc
    641    4489   42185
```

click code that checks for default which the PR skips by making the class callable : 

https://github.com/pallets/click/blob/14472ffcd80dd86d47ddc08341168152540ee6f2/src/click/core.py#L2104-L2127

Code where default is formatted

https://github.com/pallets/click/blob/14472ffcd80dd86d47ddc08341168152540ee6f2/src/click/core.py#L2755-L2771