@Bowrna 2022-04-04
@tirkarthi Can you explain to me why you have chosen to move a few imports alone to the local function? Does it depend on how heavy the import is? How do you determine that? 

It would be useful for me to understand that point.