@blag 2022-04-04
@Bowrna tirkarthi mentioned that in [this comment](https://github.com/apache/airflow/pull/22613#issuecomment-1086788871).