@tirkarthi 2022-04-05
@Bowrna Importing a package goes on to executed import statements in the package and so on. When one of the imported module does lot of work during import level like below code where toplevel has a sleep then this will delay the actual import of the file though the heavy module is used by only one function. In this case making the heavy import local will ensure the cost is paid only during function execution.

```python
# a.py
import b

def bar():
    b.foo()

if __name__ == "__main__":
    bar()
```

```python
# b.py
import c

def foo():
    pass
```

```python
# c.py
import time

time.sleep(1)
```

Now when I do measure import time you can see importing c.py (1 sec) taking a lot of time and it adds with b (201us) and then to a (1689 us) making the total time to `import a` as 1003138us (around 1s) . This is due to the fact b imports c which is costly. `b` is used only by `bar`. Now if we move the import local to function then the cost is paid only when we call bar not importing a.py

```
python -S -X importtime -c 'import a'
import time: self [us] | cumulative | imported package
import time:       224 |        224 |   _io
import time:        24 |         24 |   marshal
import time:       295 |        295 |   posix
import time:      1206 |       1747 | _frozen_importlib_external
import time:       221 |        221 |   time
import time:       171 |        391 | zipimport
import time:       101 |        101 |     _codecs
import time:       313 |        414 |   codecs
import time:       272 |        272 |   encodings.aliases
import time:       450 |       1135 | encodings
import time:       128 |        128 | encodings.utf_8
import time:        73 |         73 | _signal
import time:        76 |         76 |     _abc
import time:       557 |        632 |   abc
import time:       211 |        843 | io
import time:   1001250 |    1001250 |     c
import time:       201 |    1001450 |   b
import time:      1689 |    1003138 | a
```

```python
def bar():
    import b

    b.foo()

if __name__ == "__main__":
    bar()
```

```
python -S -X importtime -c 'import a'                  
import time: self [us] | cumulative | imported package
import time:       400 |        400 |   _io
import time:        77 |         77 |   marshal
import time:       894 |        894 |   posix
import time:      1115 |       2484 | _frozen_importlib_external
import time:       211 |        211 |   time
import time:       321 |        531 | zipimport
import time:       126 |        126 |     _codecs
import time:       369 |        494 |   codecs
import time:       282 |        282 |   encodings.aliases
import time:       987 |       1762 | encodings
import time:       639 |        639 | encodings.utf_8
import time:       110 |        110 | _signal
import time:        30 |         30 |     _abc
import time:       169 |        198 |   abc
import time:       212 |        410 | io
import time:       146 |        146 | a
```

Python also imports and caches the module in sys.modules. So further imports don't cause delay even when we call bar multiple times.

```
python -S -X importtime -c 'import a; a.bar(); a.bar()'
import time: self [us] | cumulative | imported package
import time:       136 |        136 |   _io
import time:        23 |         23 |   marshal
import time:       267 |        267 |   posix
import time:       408 |        832 | _frozen_importlib_external
import time:        96 |         96 |   time
import time:       128 |        223 | zipimport
import time:        87 |         87 |     _codecs
import time:       296 |        382 |   codecs
import time:       265 |        265 |   encodings.aliases
import time:       371 |       1017 | encodings
import time:       126 |        126 | encodings.utf_8
import time:        66 |         66 | _signal
import time:        27 |         27 |     _abc
import time:       152 |        179 |   abc
import time:       186 |        364 | io
import time:       311 |        311 | a
import time:   1001778 |    1001778 |   c
import time:       119 |    1001896 | b
```

Since starting the cli registers subcommands that import modules though one particular subcommand is used. So we don't want one command being slow making the whole cli slow. Thus moving the cost of imports to the point at which function call is made makes the cli start faster and more performant. The tradeoff is code duplication where if import is used in multiple functions then it might result in loss of little clarity compared to having toplevel imports.

Hope it helps.